maker rest validation

+++ controllers/makerRestController.js
@@ -2,6 +2,96 @@ const makerService = require('../services/MakerService.js');
 const authService = require('../services/authService.js');
 const {notifyAdmin} = require("../services/notificationService");
 
+let validatorMap = {
+    "present": async function (keysToValidate, body) {
+        let valid = {isValid: true, message: ""};
+        for (var keyString of keysToValidate) {
+            if (!body[keyString]) {
+                valid.isValid = false;
+                valid.message += `${keyString} was not valid.  `;
+            }
+        }
+        return valid;
+    },
+    "positiveIntegerOnly": async function (keysToValidate, body) {
+        let valid = {isValid: true, message: ""};
+        for (var keyString of keysToValidate) {
+            if (!body[keyString] || !Number.parseInt(body[keyString]
+                || body[keyString].includes("-") || body[keyString].includes("."))) {
+                valid.isValid = false;
+                valid.message += `${keyString} was not valid.  `;
+            }
+        }
+        return valid;
+    },
+    "noSpaces": async function (keysToValidate, body) {
+        let valid = {isValid: true, message: ""};
+        for (var keyString of keysToValidate) {
+            if (!body[keyString] || body[keyString].includes(" ")) {
+                valid.isValid = false;
+                valid.message += `${keyString} was not valid.  `;
+            }
+        }
+        return valid;
+    },
+    "positiveDecimalAllowed": async function (keysToValidate, body) {
+        let valid = {isValid: true, message: ""};
+        for (var keyString of keysToValidate) {
+            if (!body[keyString] || !Number.parseInt(body[keyString])
+                || body[keyString].includes("-")) {
+                valid.isValid = false;
+                valid.message += `${keyString} was not valid.  `;
+            }
+        }
+        return valid;
+    },
+    "decimalAllowed": async function (keysToValidate, body) {
+        let valid = {isValid: true, message: ""};
+        for (var keyString of keysToValidate) {
+            if (!body[keyString] || !Number.parseInt(body[keyString])) {
+                valid.isValid = false;
+                valid.message += `${keyString} was not valid.  `;
+            }
+        }
+        return valid;
+    },
+};
+
+/**
+ *  validates parameters
+ * @param paramArrayMap - object in form:
+ * {
+ *      present: array of string keys that should be present
+ *      positiveIntegerOnly: array of string keys that should parse to positive integers only,
+ *      noSpaces: array of string keys that should not have spaces
+ * }
+ *
+ * @param body request body to validate
+ * @returns object in the form:
+ * {
+ *      isValid: a boolean indicating whether or not the parameters were valid
+ *      message: a string description of the result
+ * }
+ */
+async function validateParams(paramArrayMap, body) {
+    let validator = {isValid: true, message: ""};
+    let paramsTypesToScan = ["present", "positiveIntegerOnly", "noSpaces", "positiveDecimalAllowed", "decimalAllowed"];
+    for (var paramName of paramsTypesToScan) {
+        let keyArray = paramArrayMap[paramName];
+        if (keyArray) {
+            let result = await validatorMap[paramName](keyArray, body);
+            if (!result.isValid) {
+                validator.isValid = false;
+                validator.message += result.message;
+            }
+        }
+    }
+    if (!validator.message) {
+        validator.message = "Valid";
+    }
+    return validator;
+}
+
 
 module.exports = {
 
@@ -38,13 +128,14 @@ module.exports = {
         res.send(onliners);
     },
 
+
     /**
      * ENDPOINT: /api/getMyRelationshipBucket
      * Retrieves the bucket minutes for the bucket associated with
      * the given relationship. Looks for values in the body in the form:
      * {
      *     "token": requester's token,
-     *     "relationshipId"1: id of the relationship with the desired plan
+     *     "relationshipId": id of the relationship with the desired plan
      * }
      * @param req
      * @param res
@@ -54,19 +145,17 @@ module.exports = {
         console.log(`Attempting to get time bucket for relationship ${req.body.relationshipId} from REST`);
         console.log(req.body);
 
-        let relationshipIdIsBad = !req.body.relationshipId || Number.parseInt(req.body.relationshipId)
-            || req.body.relationshipId.includes(".") || req.body.relationshipId.includes("-");
-        
-        if (!req.body.token
-            || relationshipIdIsBad) {
-            let responseObject = {error: "Bad request", code: 400, details: ""};
-            if (!req.body.token) {
-                responseObject.details += "token was invalid.  "
-            }
-            if (relationshipIdIsBad) {
-                responseObject.details += "relationshipId was invalid."
-            }
-            res.status(400).send(responseObject);
+        let validationResult = await validateParams(
+            {
+                "present": ["token"],
+                "positiveIntegerOnly": ["relationshipId"],
+                "noSpaces": [],
+                "positiveDecimalAllowed": [],
+                "decimalAllowed": []
+            }, req.body);
+        if (!validationResult.isValid) {
+            res.status(400).send({error: "Bad Request", code: 400, details: validationResult.message});
+            notifyAdmin({error: "Bad Request", code: 400, details: validationResult.message});
         } else {
             let email = await authService.getEmailFromToken(req.body.token).catch(err => {
                 console.log(err);
@@ -98,8 +187,18 @@ module.exports = {
     getAllMyRelationships: async (req, res) => {
         console.log(`Attempting to get all relationships for maker with token...\n${req.body.token}\n... from REST`);
         console.log(req.body);
-        if (!req.body.token) {
-            res.status(400).send({error: "Bad request", code: 400, details: "token was invalid"});
+
+        let validationResult = await validateParams(
+            {
+                "present": ["token"],
+                "positiveIntegerOnly": [],
+                "noSpaces": [],
+                "positiveDecimalAllowed": [],
+                "decimalAllowed": []
+            }, req.body);
+        if (!validationResult.isValid) {
+            res.status(400).send({error: "Bad Request", code: 400, details: validationResult.message});
+            notifyAdmin({error: "Bad Request", code: 400, details: validationResult.message});
         } else {
             let email = await authService.getEmailFromToken(req.body.token).catch(err => {
                 console.log(err);
@@ -133,19 +232,17 @@ module.exports = {
             ` with token...\n${req.body.token}\n...with relationship id ${req.body.relationshipId} from REST`);
         console.log(req.body);
 
-        let relationshipIdIsBad = !req.body.relationshipId || Number.parseInt(req.body.relationshipId)
-            || req.body.relationshipId.includes(".") || req.body.relationshipId.includes("-");
-
-        if (!req.body.token
-            || relationshipIdIsBad) {
-            let responseObject = {error: "Bad request", code: 400, details: ""};
-            if (!req.body.token) {
-                responseObject.details += "token was invalid.  "
-            }
-            if (relationshipIdIsBad) {
-                responseObject.details += "relationshipId was invalid."
-            }
-            res.status(400).send(responseObject);
+        let validationResult = await validateParams(
+            {
+                "present": ["token"],
+                "positiveIntegerOnly": ["relationshipId"],
+                "noSpaces": [],
+                "positiveDecimalAllowed": [],
+                "decimalAllowed": []
+            }, req.body);
+        if (!validationResult.isValid) {
+            res.status(400).send({error: "Bad Request", code: 400, details: validationResult.message});
+            notifyAdmin({error: "Bad Request", code: 400, details: validationResult.message});
         } else {
             let email = await authService.getEmailFromToken(req.body.token).catch(err => {
                 console.log(err);
@@ -218,8 +315,17 @@ module.exports = {
         console.log("Attempting to get maker by ID from REST: ");
         console.log(req.body);
 
-        if (!req.body.id) {
-            res.status(400).send({error: "Bad request", code: 400, details: "id was invalid"});
+        let validationResult = await validateParams(
+            {
+                "present": [],
+                "positiveIntegerOnly": ["id"],
+                "noSpaces": [],
+                "positiveDecimalAllowed": [],
+                "decimalAllowed": []
+            }, req.body);
+        if (!validationResult.isValid) {
+            res.status(400).send({error: "Bad Request", code: 400, details: validationResult.message});
+            notifyAdmin({error: "Bad Request", code: 400, details: validationResult.message});
         } else {
             let id = req.body.id;
             let result = await makerService.getMakerById(id).catch(err => {
@@ -250,8 +356,17 @@ module.exports = {
         console.log("Attempting to get maker by ID from REST: ");
         console.log(req.body);
 
-        if (!req.body.token) {
-            res.status(400).send({error: "Bad request", code: 400, details: "token was invalid"});
+        let validationResult = await validateParams(
+            {
+                "present": ["token"],
+                "positiveIntegerOnly": [],
+                "noSpaces": [],
+                "positiveDecimalAllowed": [],
+                "decimalAllowed": []
+            }, req.body);
+        if (!validationResult.isValid) {
+            res.status(400).send({error: "Bad Request", code: 400, details: validationResult.message});
+            notifyAdmin({error: "Bad Request", code: 400, details: validationResult.message});
         } else {
             let email = await authService.getEmailFromToken(req.body.token).catch(err => {
                 console.log(err);
@@ -285,10 +400,24 @@ module.exports = {
     getClientsForMaker: async (req, res) => {
         console.log("Getting client list for maker from REST");
         console.log(req.body);
-        res.send(await makerService.getClientListForMakerId(req.body.id).catch(err => {
-            console.log(err);
-            notifyAdmin(err.toString());
-        }));
+
+        let validationResult = await validateParams(
+            {
+                "present": [],
+                "positiveIntegerOnly": ["id"],
+                "noSpaces": [],
+                "positiveDecimalAllowed": [],
+                "decimalAllowed": []
+            }, req.body);
+        if (!validationResult.isValid) {
+            res.status(400).send({error: "Bad Request", code: 400, details: validationResult.message});
+            notifyAdmin({error: "Bad Request", code: 400, details: validationResult.message});
+        } else {
+            res.send(await makerService.getClientListForMakerId(req.body.id).catch(err => {
+                console.log(err);
+                notifyAdmin(err.toString());
+            }));
+        }
     },
 
 
@@ -310,18 +439,32 @@ module.exports = {
     getMyClients: async (req, res) => {
         console.log("Getting client list for maker from REST");
         console.log(req.body);
-        let email = await authService.getEmailFromToken(req.body.token).catch(err => {
-            console.log(err);
-            notifyAdmin(err.toString());
-        });
-        let makerId = await makerService.getMakerIdByEmail(email).catch(err => {
-            console.log(err);
-            notifyAdmin(err.toString());
-        });
-        res.send(await makerService.getClientListForMakerId(makerId).catch(err => {
-            console.log(err);
-            notifyAdmin(err.toString());
-        }));
+
+        let validationResult = await validateParams(
+            {
+                "present": ["token"],
+                "positiveIntegerOnly": [],
+                "noSpaces": [],
+                "positiveDecimalAllowed": [],
+                "decimalAllowed": []
+            }, req.body);
+        if (!validationResult.isValid) {
+            res.status(400).send({error: "Bad Request", code: 400, details: validationResult.message});
+            notifyAdmin({error: "Bad Request", code: 400, details: validationResult.message});
+        } else {
+            let email = await authService.getEmailFromToken(req.body.token).catch(err => {
+                console.log(err);
+                notifyAdmin(err.toString());
+            });
+            let makerId = await makerService.getMakerIdByEmail(email).catch(err => {
+                console.log(err);
+                notifyAdmin(err.toString());
+            });
+            res.send(await makerService.getClientListForMakerId(makerId).catch(err => {
+                console.log(err);
+                notifyAdmin(err.toString());
+            }));
+        }
     },
 
     /**
@@ -338,18 +481,32 @@ module.exports = {
     getMyTimeSheets: async (req, res) => {
         console.log(`Maker with token...\n${req.body.token}\n...is requesting their timesheets from REST`);
         console.log(req.body);
-        let email = await authService.getEmailFromToken(req.body.token).catch(err => {
-            console.log(err);
-            notifyAdmin(err.toString());
-        });
-        let id = await makerService.getMakerIdByEmail(email).catch(err => {
-            console.log(err);
-            notifyAdmin(err.toString());
-        });
-        res.send(await makerService.getSheetsByMaker(id).catch(err => {
-            console.log(err);
-            notifyAdmin(err.toString());
-        }));
+
+        let validationResult = await validateParams(
+            {
+                "present": ["token"],
+                "positiveIntegerOnly": [],
+                "noSpaces": [],
+                "positiveDecimalAllowed": [],
+                "decimalAllowed": []
+            }, req.body);
+        if (!validationResult.isValid) {
+            res.status(400).send({error: "Bad Request", code: 400, details: validationResult.message});
+            notifyAdmin({error: "Bad Request", code: 400, details: validationResult.message});
+        } else {
+            let email = await authService.getEmailFromToken(req.body.token).catch(err => {
+                console.log(err);
+                notifyAdmin(err.toString());
+            });
+            let id = await makerService.getMakerIdByEmail(email).catch(err => {
+                console.log(err);
+                notifyAdmin(err.toString());
+            });
+            res.send(await makerService.getSheetsByMaker(id).catch(err => {
+                console.log(err);
+                notifyAdmin(err.toString());
+            }));
+        }
     },
 
     /**
@@ -376,13 +533,27 @@ module.exports = {
     createMaker: async (req, res) => {
         console.log("Attempting to create a maker from REST: ");
         console.log(req.body);
-        let newMaker = await makerService.createNewMaker(req.body.firstName, req.body.lastName,
-            req.body.email, req.body.unique)
-            .catch(err => {
-                console.log(err);
-                notifyAdmin(err.toString());
-            });
-        res.send(newMaker);
+
+        let validationResult = await validateParams(
+            {
+                "present": ["firstName", "lastName"],
+                "positiveIntegerOnly": [],
+                "noSpaces": ["email"],
+                "positiveDecimalAllowed": [],
+                "decimalAllowed": []
+            }, req.body);
+        if (!validationResult.isValid) {
+            res.status(400).send({error: "Bad Request", code: 400, details: validationResult.message});
+            notifyAdmin({error: "Bad Request", code: 400, details: validationResult.message});
+        } else {
+            let newMaker = await makerService.createNewMaker(req.body.firstName, req.body.lastName,
+                req.body.email, req.body.unique)
+                .catch(err => {
+                    console.log(err);
+                    notifyAdmin(err.toString());
+                });
+            res.send(newMaker);
+        }
     },
 
     /**
@@ -409,13 +580,27 @@ module.exports = {
     updateMaker: async (req, res) => {
         console.log("Attempting to update a maker from REST: ");
         console.log(req.body);
-        let maker = await makerService.updateMaker(req.body.id, req.body.firstName,
-            req.body.lastName, req.body.email, req.body.unique)
-            .catch(err => {
-                console.log(err);
-                notifyAdmin(err.toString());
-            });
-        res.send(maker);
+
+        let validationResult = await validateParams(
+            {
+                "present": ["firstName", "lastName"],
+                "positiveIntegerOnly": ["id"],
+                "noSpaces": ["email"],
+                "positiveDecimalAllowed": [],
+                "decimalAllowed": []
+            }, req.body);
+        if (!validationResult.isValid) {
+            res.status(400).send({error: "Bad Request", code: 400, details: validationResult.message});
+            notifyAdmin({error: "Bad Request", code: 400, details: validationResult.message});
+        } else {
+            let maker = await makerService.updateMaker(req.body.id, req.body.firstName,
+                req.body.lastName, req.body.email, req.body.unique)
+                .catch(err => {
+                    console.log(err);
+                    notifyAdmin(err.toString());
+                });
+            res.send(maker);
+        }
     },
 
     /**
@@ -431,7 +616,21 @@ module.exports = {
     deleteMaker: async (req, res) => {
         console.log("Attempting to delete a maker from REST: ");
         console.log(req.body);
-        makerService.deleteMaker(req.body.id);
-        res.send({});
+
+        let validationResult = await validateParams(
+            {
+                "present": [],
+                "positiveIntegerOnly": ["id"],
+                "noSpaces": [],
+                "positiveDecimalAllowed": [],
+                "decimalAllowed": []
+            }, req.body);
+        if (!validationResult.isValid) {
+            res.status(400).send({error: "Bad Request", code: 400, details: validationResult.message});
+            notifyAdmin({error: "Bad Request", code: 400, details: validationResult.message});
+        } else {
+            makerService.deleteMaker(req.body.id);
+            res.send({});
+        }
     }
 };
